requirements.txt

django
Djangorestframework
psycopg2
drf_spectacular
pillow
django-cors-headers
django-filter
qrcode 
reportlab
python-dotenv

